Make `open` independent of default encoding
In python3, default encoding of `open()` comes from `locale.getpreferredencoding()`.
And it was not `utf-8` in my case (`cp949`), which caused error while reading `WorkshopFileInfos.json` on line 83. (since the file was encoded in utf-8 and contained non-ascii characters)
So I explicitly set encoding to `utf-8` when reading non-binary text files to resolve this issue.